+++ src/main/java/com/sprint/mission/discodeit/service/file/FileMessageService.java
@@ -39,6 +39,9 @@ public class FileMessageService implements MessageService {
 
     @Override
     public Message getMessage(UUID messageId) {
+        if (messageRepository.messageExists(messageId)) {
+            throw new IllegalArgumentException("존재하지 않는 메시지ID입니다.");
+        }
         return messageRepository.findById(messageId);
     }
 
@@ -54,6 +57,9 @@ public class FileMessageService implements MessageService {
 
     @Override
     public void registerMessage(UUID channelId, String userName, String messageContent) {
+        if (channelRepository.channelExists(channelId) && userRepository.userExists(userName)) {
+            throw new IllegalArgumentException("존재하지 않는 ID입니다.");
+        }
         messageRepository.createMessage(
                 channelRepository.findById(channelId),
                 userRepository.findByName(userName),
@@ -63,11 +69,17 @@ public class FileMessageService implements MessageService {
 
     @Override
     public void updateMessage(UUID messageId, String messageContent) {
+        if (messageRepository.messageExists(messageId)) {
+            throw new IllegalArgumentException("존재하지 않는 메시지ID입니다.");
+        }
         messageRepository.updateMessage(messageId, messageContent);
     }
 
     @Override
     public void deleteMessage(UUID messageId) {
+        if (messageRepository.messageExists(messageId)) {
+            throw new IllegalArgumentException("존재하지 않는 메시지ID입니다.");
+        }
         messageRepository.deleteMessage(messageId);
     }
