+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFChannelService.java
@@ -22,6 +22,9 @@ public class JCFChannelService implements ChannelService {
 
     @Override
     public Channel getChannel(UUID channelId) {
+        if (channelRepository.channelExists(channelId)) {
+            throw new IllegalArgumentException("존재하지 않는 채널ID입니다.");
+        }
         return channelRepository.findById(channelId);
     }
 
@@ -41,11 +44,17 @@ public class JCFChannelService implements ChannelService {
 
     @Override
     public void updateChannel(UUID channelId, String channelName) {
+        if (channelRepository.channelExists(channelId)) {
+            throw new IllegalArgumentException("존재하지 않는 채널ID입니다.");
+        }
         channelRepository.updateChannel(channelId, channelName);
     }
 
     @Override
     public void deleteChannel(UUID channelId) {
+        if (channelRepository.channelExists(channelId)) {
+            throw new IllegalArgumentException("존재하지 않는 채널ID입니다.");
+        }
         channelRepository.deleteChannel(channelId);
         messageRepository.deleteMessagesByChannelId(channelId);
     }
